WA-238 Adding fetchTransactions action
