+++ src/routes/safe/component/Transactions/transactions.js
@@ -1,8 +1,8 @@
 // @flow
-import { List, Map } from 'immutable'
-import { type Owner, makeOwner } from '~/routes/safe/store/model/owner'
+import { List } from 'immutable'
+import { type Owner } from '~/routes/safe/store/model/owner'
 import { load, TX_KEY } from '~/utils/localStorage'
-import { type Confirmation, type ConfirmationProps, makeConfirmation } from '~/routes/safe/store/model/confirmation'
+import { type Confirmation, makeConfirmation } from '~/routes/safe/store/model/confirmation'
 import { makeTransaction, type Transaction, type TransactionProps } from '~/routes/safe/store/model/transaction'
 
 const buildConfirmationsFrom = (owners: List<Owner>, creator: string): List<Confirmation> => {
@@ -51,25 +51,4 @@ export const createTransaction = (
   safeTransactions[safeAddress] = txsRecord.push(transaction)
 
   localStorage.setItem(TX_KEY, JSON.stringify(safeTransactions))
-}
-
-export const loadSafeTransactions = () => {
-  const safes = load(TX_KEY) || {}
-
-  return Map().withMutations((map: Map<string, List<Confirmation>>) =>
-    Object.keys(safes).map((safe: string) => {
-      const safeTxs = safes[safe]
-      const safeTxsRecord = safeTxs.map((tx: TransactionProps) => {
-        const { confirmations } = tx
-        const txRecord = makeTransaction({
-          ...tx,
-          confirmations: List(confirmations.map((conf: ConfirmationProps) =>
-            makeConfirmation({ ...conf, owner: makeOwner(conf.owner) }))),
-        })
-
-        return txRecord
-      })
-
-      return map.set(safe, List(safeTxsRecord))
-    }))
 }
